Pipelines API

+++ Header/config/scenes.hpp
@@ -7,7 +7,7 @@ namespace dg
 {
   namespace config
   {
-    enum class Scenes : uint16_t
+    enum class Scenes : uint32_t
     {
       MAIN = 0
     };
